+++ src/Adapter/SQL/MainTable.php
@@ -11,6 +11,7 @@ use Formal\ORM\{
     Specification\Property,
     Specification\Entity,
     Specification\Child,
+    Specification\Just,
 };
 use Formal\AccessLayer\{
     Table,
@@ -154,7 +155,7 @@ final class MainTable
     {
         return Column::of(
             Column\Name::of($this->definition->id()->property()),
-            Column\Type::varchar(36)->comment('UUID'),
+            Column\Type::char(36)->comment('UUID'),
         );
     }
 
@@ -268,9 +269,12 @@ final class MainTable
     /**
      * @internal
      */
-    public function delete(): Delete
+    public function delete(Specification $specification = null): Delete
     {
-        return $this->delete;
+        return match ($specification) {
+            null => $this->delete,
+            default => $this->delete->where($this->where($specification)),
+        };
     }
 
     /**
@@ -355,6 +359,18 @@ final class MainTable
                     ->match(
                         static fn($collection) => $collection->where($specification->specification()),
                         static fn() => throw new \LogicException("Unkown collection '{$specification->collection()}'"),
+                    ),
+            );
+        }
+
+        if ($specification instanceof Just) {
+            return SubQuery::of(
+                \sprintf('entity.%s', $this->definition->id()->property()),
+                $this
+                    ->optional($specification->optional())
+                    ->match(
+                        static fn($optional) => $optional->where($specification->specification()),
+                        static fn() => throw new \LogicException("Unkown optional '{$specification->optional()}'"),
                     ),
             );
         }