+++ src/Adapter/SQL/MapType.php
@@ -23,11 +23,11 @@ final class MapType
             $type instanceof Type\MaybeType => $this($type->inner())->nullable(),
             $type instanceof Type\BoolType => Table\Column\Type::tinyint(1)
                 ->comment('Boolean'),
-            $type instanceof Type\IdType => Table\Column\Type::varchar(36)
+            $type instanceof Type\IdType => Table\Column\Type::char(36)
                 ->comment('UUID'),
             $type instanceof Type\IntType => Table\Column\Type::bigint()
                 ->comment('TODO Adjust the size depending on your use case'),
-            $type instanceof Type\PointInTimeType => Table\Column\Type::varchar(32)
+            $type instanceof Type\PointInTimeType => Table\Column\Type::char(32)
                 ->comment('Date with timezone down to the microsecond'),
             default => Table\Column\Type::longtext()
                 ->comment('TODO adjust the type depending on your use case'),